% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pBessel.R
\name{pBessel}
\alias{pBessel}
\title{pBessel: asymptotic cumulative distribution of the maximum of a h dim. 
        squared Bessel bridge}
\usage{
pBessel(tn, p)
}
\description{
input: tn (test statistic; numeric)
      p (dimension of the time series; integer)
      
output: P(t_n(X) <= tn) (numeric)
}
