\name{print.cpStat}
\alias{print.cpStat}
\title{Print method for change point statistics}
\usage{
\method{print}{cpStat}(x, ...)
}
\arguments{
\item{x}{object of the class 'cpStat'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints the value of the test statistic and adds the most likely change point location.
}
\value{
Object \code{x} is return; function is called for its side effect.
}
\seealso{
\code{\link{print}}, \code{\link{wilcox_stat}}, \code{\link{CUSUM}}, \code{\link{HodgesLehmann}},
}


\author{
Sheila \enc{Görz}{Goerz}
}
