\docType{package}
\name{robcp-package}
\alias{robcp}
\alias{robcp-package}
\title{robcp: Robust Change-Point Tests}
\description{
HIER KANN EINE GENAUERE BESCHREIBUNG DES PAKETS REIN. MAN KANN Z.B. AUCH FUNKTIONEN VERLINKEN: \code{\link{huber_cusum}}.
}

\seealso{
...
}


\author{
\strong{Maintainer}: Sheila \enc{Görz}{Goerz} \email{sheila.goerz@tu-dortmund.de} 
}