\name{zeros}
\alias{zeros}
\docType{data}
\title{
Zero of the Bessel function of first kind
}
\description{
Contains the zeros of the Bessel function of first kind.
}
\usage{data("zeros")}
\format{
  A data frame where the i-th column contains the first 50 zeros of the Bessel 
  function of the first kind and ((i - 1) / 2)th order, i = 1, ..., 5001.
}
\source{
The zeros are computed by the mathematical software octave.
}
\references{
Eaton, J., Bateman, D., Hauberg, S., Wehbring, R. (2015). "GNU Octave version 4.0.0 manual: a high-level interactive language for numerical computations".
}
\keyword{datasets}
