\name{print.cpStat}
\alias{print.cpStat}
\title{Print method for change point statistics}
\usage{
\method{print}{cpStat}(x, ...)
}
\arguments{
\item{x}{object of the class 'cpStat'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints the value of the test statistic and add the most likely change point location.
}
\seealso{
\code{\link{print}}, \code{\link{wilcox_stat}}, \code{\link{CUSUM}}, \code{\link{HodgesLehmann}},
}
